route id

+++ application/controllers/rest/Equipe.php
@@ -1,22 +1,33 @@
 <?php
 defined('BASEPATH') or exit('No direct script access allowed');
+
 use Restserver\Libraries\REST_Controller;
 use Restserver\Libraries\REST_Controller_Definitions;
+
 require APPPATH . '/libraries/REST_Controller.php';
 require APPPATH . '/libraries/REST_Controller_Definitions.php';
 require APPPATH . '/libraries/Format.php';
 
-class Equipe extends REST_Controller {
-    public function __construct() {
+class Equipe extends REST_Controller
+{
+    public function __construct()
+    {
         parent::__construct();
-        $this->load->model('Equipe_model','eq');
+        $this->load->model('Equipe_model', 'eq');
     }
-    public function equipe_get() {
-        $data = $this->eq->get();
-        $this->set_response($data,REST_Controller_Definitions::HTTP_OK);
+    public function index_get()
+    {
+        $id = (int) $this->get('id');
+        if ($id <= 0) {
+            $data = $this->eq->get();
+        } else {
+            $data = $this->eq->getOne($id);
+        }
+        $this->set_response($data, REST_Controller_Definitions::HTTP_OK);
     }
-    public function equipe_post() {
-        if((!$this->post('nome'))) {
+    public function index_post()
+    {
+        if ((!$this->post('nome'))) {
             $this->set_response([
                 'status' => false,
                 'error' => 'Campo não preenchidos'
@@ -26,7 +37,7 @@ class Equipe extends REST_Controller {
         $data = array(
             'nome' => $this->post('nome')
         );
-        if($this->eq->insert($data)) {
+        if ($this->eq->insert($data)) {
             $this->set_response([
                 'status' => true,
                 'message' => 'Equipe inserido com successo !'
@@ -38,8 +49,17 @@ class Equipe extends REST_Controller {
             ], REST_Controller_Definitions::HTTP_BAD_REQUEST);
         }
     }
-    public function equipe_delete($id) {
-        if($this->eq->delete($id)) {
+    public function index_delete()
+    {
+        $id = (int) $this->get('id');
+        if ($id <= 0) {
+            $this->set_response([
+                'status' => false,
+                'error' => 'Parâmetros obrigatórios não fornecidos'
+            ], REST_Controller_Definitions::HTTP_BAD_REQUEST);
+            return;
+        }
+        if ($this->eq->delete($id)) {
             $this->set_response([
                 'status' => true,
                 'message' => 'Equipe deletado com successo !'
@@ -51,8 +71,9 @@ class Equipe extends REST_Controller {
             ], REST_Controller_Definitions::HTTP_BAD_REQUEST);
         }
     }
-    public function equipe_put($id) {
-        if ((!$this->put('nome'))) {
+    public function index_put() {
+        $id = (int) $this->get('id');
+        if ((!$this->put('nome')) || ($id <= 0)) {
             $this->set_response([
                 'status' => false,
                 'error' => 'Campo não preenchidos'
@@ -75,4 +96,3 @@ class Equipe extends REST_Controller {
         }
     }
 }
-?>
